refactor: split homepage customer proof carousel
